+++ doc/user/content/release-notes.md
@@ -48,6 +48,10 @@ Wrap your release notes at the 80 character mark.
 
 {{% version-header v0.8.4 %}}
 
+- Change the type of the [`mz_metrics`](/sql/system-catalog#mz_metrics).`time`
+  column from [`timestamp`] to [`timestamp with time zone`] to better reflect
+  that the timestamp is in UTC.
+
 - **Breaking change.** Reject [Protobuf sources] whose schemas contain
   unsigned integer types (`uint32`, `uint64`, `fixed32`, and `fixed64`).
   Materialize previously converted these types to